+++ neoforge21/src/main/java/com/vecoo/legendcontrol/config/StorageConfig.java
@@ -20,6 +20,7 @@ public class StorageConfig extends AbstractYamlConfig {
     private long keepAliveTime = 60000L;
     private long connectionTimeout = 5000L;
     private boolean useSsl = true;
+    private int threadPool = 5;
 
     public String getStorageType() {
         return this.storageType;
@@ -71,5 +72,9 @@ public class StorageConfig extends AbstractYamlConfig {
 
     public boolean isUseSSL() {
         return this.useSsl;
+    }
+
+    public int getThreadPool() {
+        return this.threadPool;
     }
 }