+++ changelog.md
@@ -1,3 +1,3 @@
-* Fix storage module defender.
-* Fix catch conflict module defender.
-* Various improvements and fixes.
+* Fix database storage.
+* New database settings.
+* Fix discord webhook.